refactor(fgumi-sort): rename raw/inline_buffer/raw_bam_reader for clarity
