+++ crates/fgumi-sort/src/reader.rs
